shareGlobs
By writing globs into the directory paths, this repo can be used as a
workshop by any contributor to Crosswalks.It allows for the process
steps required to analyze metadata for completeness with respect to a
recommendation to be carried out in a branch, leaving a complete record
(minus the Crosswalks black box). Improvements to processing can be
shared among the group and specific problems can be shared in an easy
to replicate fashion.