OneStepCloser
In a continued attempt to move the working workflow from sole expert
use into an automated process, scripts are being altered to fit into
the process.

Currently the master works, but it would be best if there was one
script that ran all of the other scripts without lists of
collectionNames being supplied.

Current goal is something like:

./runWorkflow DataOne FGDC EML,CSDGM,Dryad,Onedcx,BDP,Mercury

One problem with this is that runTransform needs the collectionName as
well as the above arguments. fixing this will allow the dialectCodes to
be the only multiple processing.