metadataCleanRubricRigged_reliableResults
To accommodate variations within dialects alterations such as removing
namespace prefixes from rubrics and records. Minor revisions to rubric
ISO-1_EML. EML as it is written in AllCrosswalks does not have any
dialect definitions for the concepts in the Conditional component of
the ISO-1 recommendation.

This commit also contains a spreadsheet that has accurate assessments
of the contents of the DataOne Sample Set, as the dialect stands in
AllCrosswalks.

The next step is to replicate the analysis in the different dialects so
that I can get initial reports for as many salient concepts to ensure
quality of dialect definitions for concepts to improve quality of
analysis. Will also spend some time attempting to automate
visualizations for the reports with a focus on member node instead of
dialect to accommodate the needs of organizations like DataOne.