cleaningBranch
Need a template directory that allows me to churn recommendations. This
required removal of rubrics, json, and the report as well as some
directory structure.

This leaves me with the task of rubric creation and editing the
resultant xsl in the ways that I altered the ones in the ISO-1 branch