integratingUSGScollections
In preparation for Friday’s lab, I’ve uploaded the USGS metadata
received from Sky Bristol in 2015 and Madison Langseth in 2016.
Alterations include conforming the directory structure to the directory
structure of this repo, and removing the uuid folders in the old
ScienceBase created by the xsl extraction of the initial OAI-PMH pull.